Fix VTS crashes for VtsHalNeuralnetworksV1_2Target
Since Bool types are represented as uint8_t in nnApi, fixed the
incorrect handling of bool type.
Handled exceptions raised while calling IE APIs for Network
creation and serialization.
Additionally, Fixed misplaced LogTags.

+++ ngraph_creator/operations/src/Conv_2d.cpp
@@ -111,7 +111,7 @@ std::shared_ptr<ngraph::Node> Conv_2d::createNode() {
                 dilation_width_factor =
                     sModelInfo->ParseOperationInput<uint32_t>(mNnapiOperationIndex, 11);
             case 11:
-                layout = sModelInfo->ParseOperationInput<uint32_t>(mNnapiOperationIndex, 10);
+                layout = sModelInfo->ParseOperationInput<uint8_t>(mNnapiOperationIndex, 10);
             default:
                 break;
         }
@@ -136,7 +136,7 @@ std::shared_ptr<ngraph::Node> Conv_2d::createNode() {
                 dilation_width_factor =
                     sModelInfo->ParseOperationInput<uint32_t>(mNnapiOperationIndex, 8);
             case 8:
-                layout = sModelInfo->ParseOperationInput<uint32_t>(mNnapiOperationIndex, 7);
+                layout = sModelInfo->ParseOperationInput<uint8_t>(mNnapiOperationIndex, 7);
             default:
                 break;
         }
